requirements.txt
tabulate
asyncpg~=0.30.0
certifi~=2025.10.5
pyyaml~=6.0.3
python-dotenv~=1.2.1
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
ruff==0.14.4
discord.py
motor~=3.7.1
pymongo>=4.4.0,<5.0.0
